+++ app/helpers/application_helper.rb
@@ -1,2 +1,5 @@
 module ApplicationHelper
+  def simple_check(checked)
+    checked ? '<span class="glyphicon glyphicon-ok"></span>'.html_safe : '<span class="glyphicon glyphicon-unchecked"></span>'.html_safe
+  end
 end