+++ app/helpers/builder/answer_options_helper.rb
@@ -1,0 +1,2 @@
+module Builder::AnswerOptionsHelper
+end