+++ app/helpers/builder/builder_helper.rb
@@ -1,0 +1,2 @@
+module Builder::BuilderHelper
+end