+++ app/helpers/builder/questions_helper.rb
@@ -1,0 +1,2 @@
+module Builder::QuestionsHelper
+end