+++ app/models/answer_options_answer_template.rb
@@ -1,4 +1,0 @@
-class AnswerOptionsAnswerTemplate < ActiveRecord::Base
-  belongs_to :answer_template, -> { where deleted: false }
-  belongs_to :answer_option, -> { where deleted: false }
-end