+++ app/models/answer_templates_question.rb
@@ -1,4 +1,0 @@
-class AnswerTemplatesQuestion < ActiveRecord::Base
-  belongs_to :answer_template, -> { where deleted: false }
-  belongs_to :question, -> { where deleted: false }
-end