+++ app/models/question.rb
@@ -6,11 +6,12 @@ class Question < ActiveRecord::Base
   # Attribute related macros
   # Associations
   belongs_to :user
+  belongs_to :survey
   has_many :answer_templates, -> { current.order("answer_templates.created_at asc") }
   has_many :answers
 
   # Validations
-  validates_presence_of :text, :slug, :user_id
+  validates_presence_of :title, :slug, :user_id
   validates_uniqueness_of :slug, scope: [ :deleted ]
   validates_format_of :slug, with: /\A(?!\Anew\Z)[a-z][a-z0-9\-]*\Z/