+++ config/routes.rb
@@ -1,15 +1,23 @@
 Rails.application.routes.draw do
 
   resources :surveys
+
+  namespace :builder do
+    resources :surveys, shallow: true do
+      resources :questions do
+        resources :answer_templates do
+          resources :answer_options
+        end
+      end
+    end
+
+    get '', to: redirect("builder/surveys")
+  end
+
   resources :research_topics do
     resources :comments
     resources :votes
   end
-  get 'members', to: 'members#index'
-  get 'members/:forum_name', to: 'members#show', as: 'member'
-  get 'leaderboards', to: 'members#leaderboards'
-  get 'settings', to: 'account#settings'
-  patch 'settings', to: 'account#update', as: 'update_settings'
 
   resources :forums do
     resources :topics do
@@ -20,6 +28,13 @@ Rails.application.routes.draw do
       end
     end
   end
+
+  get 'members', to: 'members#index'
+  get 'members/:forum_name', to: 'members#show', as: 'member'
+  get 'leaderboards', to: 'members#leaderboards'
+  get 'settings', to: 'account#settings'
+  patch 'settings', to: 'account#update', as: 'update_settings'
+
   scope module: 'application' do
     get :dashboard
     get :version