+++ db/migrate/20150930194607_drop_answer_options_answer_templates.rb
@@ -1,0 +1,5 @@
+class DropAnswerOptionsAnswerTemplates < ActiveRecord::Migration
+  def change
+    drop_table :answer_options_answer_templates
+  end
+end