+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 1.0.0
 
 - Skeleton files to initialize Rails application with testing framework and continuous integration
-- Functional forum template
-- Functional research topic template
-- Survey model built
+- Functional forum created
+- Functional research topic ranking created
+- Functional survey builder created